requirements.txt
streamlit
numpy
pandas
google-generativeai 